+++ arm_inlines_linearization.h
@@ -10,6 +10,8 @@
 #ifdef __cplusplus
 	extern "C" {
 #endif
+
+#include "arm_typedefs.h"			// Common typedefs
 /****************************************************************/