+++ arm_inlines_var_size.h
@@ -10,6 +10,9 @@
 #ifdef __cplusplus
 	extern "C" {
 #endif
+
+#include "arm_macros.h"				// Common macros
+#include "arm_typedefs.h"			// Common typedefs
 /****************************************************************/