+++ arm_macros.h
@@ -11,16 +11,15 @@
 	extern "C" {
 #endif
 
-#include <math.h>				// For math constants
-#include "arm_typedefs.h"		// Common typedefs
+#include "arm_typedefs.h"			// Common typedefs
 /****************************************************************/
 
 
-#define	Undefined			-1				//!< Undefined value
+#define	Undefined			(-1)			//!< Undefined value
 
-#define Null				0				//!< Null Value
-#define pNull				(void *) 0		//!< Null pointer -> same as NULL in stdlib.h
-#define	charNUL				'\0'			//!< Null Char
+#define Null				(0)				//!< Null Value
+#define pNull				((void *) 0)	//!< Null pointer -> same as NULL in stdlib.h
+#define	charNull			'\0'			//!< Null Char
 
 
 #ifndef True
@@ -45,6 +44,7 @@
 //!\warning this macro is optimized only when used with \b b with a static value
 #define RSHIFT64(v, b)		((v) / (1ULL << b))											//!< Shift \b v \b b bits right (up to 63b)
 
+
 #define MAKEWORD(lsb, msb)	((WORD) (((BYTE) (lsb)) | LSHIFT(((WORD) ((BYTE) (msb))), 8)))		//!< Make WORD from \b lsb and \b msb
 #define MAKELONG(lsw, msw)	((DWORD) (((WORD) (lsw)) | LSHIFT(((DWORD) ((WORD) (msw))), 16)))	//!< Make LONG from \b lsw and \b msw
 
@@ -57,18 +57,18 @@
 #undef HIBYTE	//!\note Undefine HIBYTE if already defined in some other library
 #endif
 
-#define LOBYTE(w)			((BYTE) (w))						//!< Get BYTE LSB from WORD \b w
-#define HIBYTE(w)			((BYTE) RSHIFT((WORD) (w), 8))		//!< Get BYTE MSB from WORD \b w
-#define LOWORD(l)			((WORD) (l))						//!< Get WORD LSW from LONG \b l
-#define HIWORD(l)			((WORD) RSHIFT((DWORD) (l), 16))	//!< Get WORD MSW from LONG \b l
+#define LOBYTE(w)				((BYTE) (w))						//!< Get BYTE LSB from WORD \b w
+#define HIBYTE(w)				((BYTE) RSHIFT((WORD) (w), 8))		//!< Get BYTE MSB from WORD \b w
+#define LOWORD(l)				((WORD) (l))						//!< Get WORD LSW from LONG \b l
+#define HIWORD(l)				((WORD) RSHIFT((DWORD) (l), 16))	//!< Get WORD MSW from LONG \b l
 
-#define	SWAP_TYPE(a, b, typ)	{ typ c = a; a = b; b = c; }	//!< Swap type \b typ \b a \& \b b
-#define	SWAP_BYTE(a, b)			SWAP_TYPE(a, b, BYTE)			//!< Swap BYTEs \b a \& \b b
-#define	SWAP_WORD(a, b)			SWAP_TYPE(a, b, WORD)			//!< Swap WORDs \b a \& \b b
-#define	SWAP_DWORD(a, b)		SWAP_TYPE(a, b, DWORD)			//!< Swap DWORDs \b a \& \b b
-#define	SWAP_LWORD(a, b)		SWAP_TYPE(a, b, LWORD)			//!< Swap LWORDs \b a \& \b b
-#define	SWAP_FLOAT(a, b)		SWAP_TYPE(a, b, float)			//!< Swap floats \b a \& \b b
-#define	SWAP_DOUBLE(a, b)		SWAP_TYPE(a, b, double)			//!< Swap doubles \b a \& \b b
+#define	SWAP_TYPE(a, b, typ)	{ typ c = a; a = b; b = c; }		//!< Swap type \b typ \b a \& \b b
+#define	SWAP_BYTE(a, b)			SWAP_TYPE(a, b, BYTE)				//!< Swap BYTEs \b a \& \b b
+#define	SWAP_WORD(a, b)			SWAP_TYPE(a, b, WORD)				//!< Swap WORDs \b a \& \b b
+#define	SWAP_DWORD(a, b)		SWAP_TYPE(a, b, DWORD)				//!< Swap DWORDs \b a \& \b b
+#define	SWAP_LWORD(a, b)		SWAP_TYPE(a, b, LWORD)				//!< Swap LWORDs \b a \& \b b
+#define	SWAP_FLOAT(a, b)		SWAP_TYPE(a, b, float)				//!< Swap floats \b a \& \b b
+#define	SWAP_DOUBLE(a, b)		SWAP_TYPE(a, b, double)				//!< Swap doubles \b a \& \b b
 
 #define VAL_AT(addr, typ)		(*(typ *) (addr))										//!< Get the type \b typ content of address \b addr
 
@@ -81,14 +81,31 @@
 #define OFFSET_OF(typ, mbr)		((size_t) &(((typ *)0)->mbr))							//!< Computes the offset member \b mbr from struct type \b typ
 #define ROOT_OF(ptr, typ, mbr)	((typ *) (((uint8_t *) ptr) - OFFSET_OF(typ, mbr)))		//!< Computes the address of parent struct \b typ of \b ptr from member \b mbr
 
+
 //! \warning No nesting possible, use \a XCAT in this case
 #define	CAT(a, b)			a##b			//!< Preprocessor Name catenation
 #define XCAT(a, b)			CAT(a, b)		//!< Preprocessor Name catenation (possible nesting)
 
 #define STR(s)				(#s)			//!< Stringify an expression
 
+
 #define binEval(exp)		((exp) ? true : false)		//!< boolean evaluation of expression \b exp
 #define nbinEval(exp)		(!binEval(exp))				//!< complemented boolean evaluation of expression \b exp
+
+
+#define CLAMP(v, min, max)	((v) < (min) ? (min) : ((v) > (max) ? (max) : (v)))		//!< Returns the value between \b min and \b max from \b val
+#define	MAX3(a, b, c)		((b) >= (c) ? ((a) >= (b) ? (a) : (b)) : ((a) >= (c) ? (a) : (c)))	//!< Returns max value between \b a, \b b and \b c
+#define	MIN3(a, b, c)		((b) <= (c) ? ((a) <= (b) ? (a) : (b)) : ((a) <= (c) ? (a) : (c)))	//!< Returns min value between \b a, \b b and \b c
+
+#ifndef clamp
+	#define clamp	CLAMP	//!< \b clamp alias for \b CLAMP
+#endif
+#ifndef max3
+	#define max3	MAX3	//!< \b max3 alias for \b MAX3
+#endif
+#ifndef min3
+	#define min3	MIN3	//!< \b min3 alias for \b MIN3
+#endif
 
 #ifndef max
 	#define max(a, b)		((a) >= (b) ? (a) : (b))	//!< Returns max value between \b a and \b b
@@ -97,24 +114,14 @@
 	#define min(a, b)		((a) <= (b) ? (a) : (b))	//!< Returns min value between \b a and \b b
 #endif
 
-#define	MIN3(a, b, c)		((b) <= (c) ? ((a) <= (b) ? (a) : (b)) : ((a) <= (c) ? (a) : (c)))	//!< Returns max value between \b a, \b b and \b c
-#define	MAX3(a, b, c)		((b) >= (c) ? ((a) >= (b) ? (a) : (b)) : ((a) >= (c) ? (a) : (c)))	//!< Returns min value between \b a, \b b and \b c
-
-#define CLAMP(v, min, max)	((v) < (min) ? (min) : ((v) > (max) ? (max) : (v)))		//!< Returns the value between \b min and \b max from \b val
-
-#define OneThird			((float) (1.0f / 3.0f))		//!< 1/3 approximation
-#define TwoThird			((float) (2.0f / 3.0f))		//!< 2/3 approximation
-
-#define Pi					M_PI						//!< Pi approximation alias
 
 #define BYTE_TO_PERC(b)		((BYTE) (((b) * 100) / 255))						//!< Converts a BYTE \b b (0-255) to percent (0-100)
 #define PERC_TO_BYTE(p)		((BYTE) (((p) > 100 ? 100 : (p)) * 255 / 100))		//!< Converts a BYTE \b p percentage (0-100) to BYTE (0-255) with max checking
 
-#define RAD_TO_FLOAT(r)		((float) (((r) > M_TWOPI ? M_TWOPI : (r)) / M_TWOPI))
-#define FLOAT_TO_RAD(f)		((float) ((((f) > 1.0f ? 1.0f : (f)) < 0.0f ? 0.0f : (f)) * M_TWOPI)
 
-#define DEG_TO_FLOAT(d)		((float) (((d) > 360.0f ? 360.0f : (d)) / 360.0f))
-#define FLOAT_TO_DEG(f)		((float) ((((f) > 1.0f ? 1.0f : (f)) < 0.0f ? 0.0f : (f)) * 360.0f))
+/*** Constants ***/
+#define M_1_3				(1.0 / 3.0)		//!< 1/3 constant approximation
+#define M_2_3				(2.0 / 3.0)		//!< 2/3 constant approximation
 
 
 /*** Test Macros ***/