+++ arm_stdclib.h
@@ -13,23 +13,24 @@
 
 #include <string.h>
 #include <stdio.h>
+#include "arm_macros.h"				// Common macros
 /****************************************************************/
 
 
 /*** prints and strings ***/
-#define str_clr(s)			(s[0] = '\0')						//!< clear string \b s (fast way)
-#define str_clr_safe(s)		(memset(s, '\0', sizeof(s)))		//!< clear string \b s (safe way)
+#define str_clr(s)			(s[0] = charNull)						//!< clear string \b s (fast way)
+#define str_clr_safe(s)		(memset(s, charNull, sizeof(s)))		//!< clear string \b s (safe way)
 
-#define	str_add_tab(s)		(strcat(s, '\t'))					//!< Adding tab to string using strcat
-#define	str_add_cr(s)		(strcat(s, '\r\n'))					//!< Adding new line to string using strcat
+#define	str_add_tab(s)		(strcat(s, '\t'))						//!< Adding tab to string using strcat
+#define	str_add_cr(s)		(strcat(s, '\r\n'))						//!< Adding new line to string using strcat
 
 
-#define printExpr(e)		(printf("%s = %d\r\n", #e, (e)))	//!< Print expression \b e and it's result \b e using printf
+#define printExpr(e)		(printf("%s = %d\r\n", STR(e), (e)))	//!< Print expression \b e and it's result \b e using printf
 
-#define verbInstr(i)		(printf(#i), (i))					//!< Print instruction \b e and execute it
+#define verbInstr(i)		(printf(STR(i)), (i))					//!< Print instruction \b e and execute it
 
-#define verbInc(x)			(puts("Incrementing " #x), (x)++)	//!< Increment example using puts
-#define verbDec(x)			(puts("Decrementing " #x), (x)--)	//!< Decrement example using puts
+#define verbInc(x)			(puts("Incrementing " STR(x)), (x)++)	//!< Increment example using puts
+#define verbDec(x)			(puts("Decrementing " STR(x)), (x)--)	//!< Decrement example using puts
 
 
 /****************************************************************/