+++ arm_typedefs.h
@@ -33,7 +33,7 @@
 		defined(__ARMEB__) || defined(__THUMBEB__) || defined(__AARCH64EB__) ||					\
 		defined(_MIPSEB) || defined(__MIPSEB) || defined(__MIPSEB__)
 	#ifndef BIG_ENDIAN
-		#define BIG_ENDIAN		// Big-endian target
+		#define BIG_ENDIAN		//!< Big-endian target
 	#endif
 #elif	defined(__BYTE_ORDER) && (__BYTE_ORDER == __LITTLE_ENDIAN) ||							\
 		defined(__BYTE_ORDER__) && (__BYTE_ORDER__ == __ORDER_LITTLE_ENDIAN__) ||				\
@@ -42,12 +42,12 @@
 		defined(__ARMEL__) || defined(__THUMBEL__) || defined(__AARCH64EL__) ||					\
 		defined(_MIPSEL) || defined(__MIPSEL) || defined(__MIPSEL__)
 	#ifndef LITTLE_ENDIAN
-		#define LITTLE_ENDIAN	// Little-endian target
+		#define LITTLE_ENDIAN	//!< Little-endian target
 	#endif
 #elif	defined(__BYTE_ORDER) && (__BYTE_ORDER == __PDP_ENDIAN)
 	#warning "Unfortunately, easy-access typedefs for PDP Endian not handled by the sarmfw library!"
 	#ifndef PDP_ENDIAN
-		#define PDP_ENDIAN		// Little-endian (word swapped) target
+		#define PDP_ENDIAN		//!< Little-endian (word swapped) target
 	#endif
 #else
 	#warning "Unknown endianness, please define LITTLE_ENDIAN or BIG_ENDIAN symbol to ensure behavior!"